Code cleaning: JS deletions and file splitting is done.
