+++ FrontEnd/assets/scripts/get_works.js
@@ -1,11 +1,17 @@
+import {
+    worksURL
+} from "./script.js";
+import {
+    replaceSpaceByUnderscore
+} from "./helpers/string_replacer.js";
+
 /****** Step 1.1 get works from backend ******/
-export let categories = new Set();
 /**
  * This function fetches data and fills the localStorage for speed and less network use during next page reloads.
  * @returns : an array of fetched works in JSON format is returned.
  */
 export async function fetchAndStoreWorks() {
-    fetch("http://localhost:5678/api/works").then(works => works.json()).then(worksJSON => {
+    fetch(worksURL).then(works => works.json()).then(worksJSON => {
         if (Array.isArray(worksJSON)) {
             window.localStorage.setItem("works", JSON.stringify(worksJSON));
             return worksJSON;
@@ -21,16 +27,18 @@ export async function fetchAndStoreWorks() {
 
 /**
  * This function creates HTML elements in <div class="gallery"> based on works from the API.
- * @param {Promise<any>} works : works in JSON format is expected from the API at promise resolution
+ * It is called in script.js line 55.
+ * @param {Promise<any>} worksPromise : works in JSON format is expected from the API at promise resolution
  * @param {Element} galleryDiv : the <div class="gallery"> including figures
  * @param {HTMLElement[]} initialFetchedGallery : a copy of the gallery initially fetched from the API
  * @returns : the array of figures initially fetched from the API
 */
-export async function fillGallery(works, galleryDiv, initialFetchedGallery) {
+export async function fillGallery(worksPromise, galleryDiv, initialFetchedGallery) {
     try{
-        const result = await works;
+        const works = await worksPromise;
         const figures = [];
-        result.forEach(work => {
+
+        works.forEach(work => {
             let imgFromAPI = document.createElement("img");
             const titleFromAPI = work.title;
             imgFromAPI.src = work.imageUrl;
@@ -42,7 +50,6 @@ export async function fillGallery(works, galleryDiv, initialFetchedGallery) {
             let figureFromAPI = document.createElement("figure");
 
             let categ = work.category.name;
-            categories.add(categ);
             categ = replaceSpaceByUnderscore(categ);
             figureFromAPI.classList.add(categ);
 
@@ -60,17 +67,4 @@ export async function fillGallery(works, galleryDiv, initialFetchedGallery) {
     } catch(error) {
         console.error("Error at filling the gallery: ", error);
     }
-}
-
-/**
- * This function replaces spaces by underscores ("_").
- * @param {string} name : the class name including one or more spaces (" ")
- * @returns the string substitution with "_" instead of " "
- */
-export function replaceSpaceByUnderscore(name) {
-    let underscored = name;
-    if(name.includes(" ")) {
-        underscored = name.replaceAll(" ", "_");
-    }
-    return underscored;
 }
